Fix the swapchain's resolution issue
The surface object's extent can be smaller than the asked resolution. Like if the window is not full and if you try to ask for the full resolution of your monitor, the height might be a bit smaller, which will cause a vulkan validation error. So, using the surface's extent to initialise the swapchain.