Fix the AddMultiple function
Since I had changed the Reserving logic, the AddMultiple function had been broken but because the function was marked noexcept, it was silently crashing the test units and I hadn't noticed.

This commit will break the compilation because I found out about this after making another huge change. So, have to do this.