Move the Compute PSO creation
In the SetShaderPath function. The Graphics PSOs are created on demand when the Fragment Shader function is called or a new Model asks for a certain fragment shader.

But the compute ones won't have multiple variations. But it can't be created in the ctor as the ShaderPath won't be set then.